classify-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
open-clip-torch==2.26.1
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
scikit-learn==1.5.2
numpy==1.26.4
sqlalchemy==2.0.35
aiofiles==24.1.0
jinja2==3.1.4
httpx==0.27.2
